Language.C.Clang.Cursor

Description

If you know what CursorKinds you want to operate on, consider using Language.C.Clang.Cursor.Typed instead.

Documentation

data Cursor

translationUnitCursor :: TranslationUnit -> Cursor

cursorTranslationUnit :: Cursor -> TranslationUnit

cursorChildrenF :: Fold Cursor Cursor

Fold over the children of a cursor in the lens sense.

cursorChildren :: Cursor -> [Cursor]

cursorDescendantsF :: Fold Cursor Cursor

Fold over a Cursor and all of its descendants recursively.

cursorDescendants :: Cursor -> [Cursor]

List a Cursor and all of its descendants recursively.

cursorSpelling :: Cursor -> ByteString

cursorExtent :: Cursor -> Maybe SourceRange

cursorUSR :: Cursor -> ByteString

cursorReferenced :: Cursor -> Maybe Cursor

cursorType :: Cursor -> Maybe Type

cursorKind :: Cursor -> CursorKind

data TypeLayoutError

Constructors

TypeLayoutErrorInvalid

TypeLayoutErrorIncomplete

TypeLayoutErrorDependent

offsetOfField :: Cursor -> Either TypeLayoutError Word64

data CursorKind

Constructors

UnexposedDecl

StructDecl

UnionDecl

ClassDecl

EnumDecl

FieldDecl

EnumConstantDecl

FunctionDecl

VarDecl

ParmDecl

ObjCInterfaceDecl

ObjCCategoryDecl

ObjCProtocolDecl

ObjCPropertyDecl

ObjCIvarDecl

ObjCInstanceMethodDecl

ObjCClassMethodDecl

ObjCImplementationDecl

ObjCCategoryImplDecl

TypedefDecl

CXXMethod

Namespace

LinkageSpec

Constructor

Destructor

ConversionFunction

TemplateTypeParameter

NonTypeTemplateParameter

TemplateTemplateParameter

FunctionTemplate

ClassTemplate

ClassTemplatePartialSpecialization

NamespaceAlias

UsingDirective

UsingDeclaration

TypeAliasDecl

ObjCSynthesizeDecl

ObjCDynamicDecl

CXXAccessSpecifier

FirstDecl

LastDecl

FirstRef

ObjCSuperClassRef

ObjCProtocolRef

ObjCClassRef

TypeRef

CXXBaseSpecifier

TemplateRef

NamespaceRef

MemberRef

LabelRef

OverloadedDeclRef

VariableRef

LastRef

FirstInvalid

InvalidFile

NoDeclFound

NotImplemented

InvalidCode

LastInvalid

FirstExpr

UnexposedExpr

DeclRefExpr

MemberRefExpr

CallExpr

ObjCMessageExpr

BlockExpr

IntegerLiteral

FloatingLiteral

ImaginaryLiteral

StringLiteral

CharacterLiteral

ParenExpr

UnaryOperator

ArraySubscriptExpr

BinaryOperator

CompoundAssignOperator

ConditionalOperator

CStyleCastExpr

CompoundLiteralExpr

InitListExpr

AddrLabelExpr

StmtExpr

GenericSelectionExpr

GNUNullExpr

CXXStaticCastExpr

CXXDynamicCastExpr

CXXReinterpretCastExpr

CXXConstCastExpr

CXXFunctionalCastExpr

CXXTypeidExpr

CXXBoolLiteralExpr

CXXNullPtrLiteralExpr

CXXThisExpr

CXXThrowExpr

CXXNewExpr

CXXDeleteExpr

UnaryExpr

ObjCStringLiteral

ObjCEncodeExpr

ObjCSelectorExpr

ObjCProtocolExpr

ObjCBridgedCastExpr

PackExpansionExpr

SizeOfPackExpr

LambdaExpr

ObjCBoolLiteralExpr

ObjCSelfExpr

LastExpr

FirstStmt

UnexposedStmt

LabelStmt

CompoundStmt

CaseStmt

DefaultStmt

IfStmt

SwitchStmt

WhileStmt

DoStmt

ForStmt

GotoStmt

IndirectGotoStmt

ContinueStmt

BreakStmt

ReturnStmt

GCCAsmStmt

AsmStmt

ObjCAtTryStmt

ObjCAtCatchStmt

ObjCAtFinallyStmt

ObjCAtThrowStmt

ObjCAtSynchronizedStmt

ObjCAutoreleasePoolStmt

ObjCForCollectionStmt

CXXCatchStmt

CXXTryStmt

CXXForRangeStmt

SEHTryStmt

SEHExceptStmt

SEHFinallyStmt

MSAsmStmt

NullStmt

DeclStmt

OMPParallelDirective

OMPSimdDirective

OMPForDirective

OMPSectionsDirective

OMPSectionDirective

OMPSingleDirective

OMPParallelForDirective

OMPParallelSectionsDirective

OMPTaskDirective

OMPMasterDirective

OMPCriticalDirective

OMPTaskyieldDirective

OMPBarrierDirective

OMPTaskwaitDirective

OMPFlushDirective

SEHLeaveStmt

LastStmt

TranslationUnit

FirstAttr

UnexposedAttr

IBActionAttr

IBOutletAttr

IBOutletCollectionAttr

CXXFinalAttr

CXXOverrideAttr

AnnotateAttr

AsmLabelAttr

PackedAttr

PureAttr

ConstAttr

NoDuplicateAttr

CUDAConstantAttr

CUDADeviceAttr

CUDAGlobalAttr

CUDAHostAttr

LastAttr

PreprocessingDirective

MacroDefinition

MacroExpansion

MacroInstantiation

InclusionDirective

FirstPreprocessing

LastPreprocessing

ModuleImportDecl

FirstExtraDecl

LastExtraDecl